% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{admix_subset}
\alias{admix_subset}
\title{Admixture Data Subsetting}
\usage{
admix_subset(data, anc = NULL, pct = NULL, ...)
}
\arguments{
\item{data}{Data frame containing the admixture data.}

\item{anc}{Vector of ancestry column names to use for pairwise subsetting with percentage vector. Must be of same length as the supplied percentage vector.}

\item{pct}{Vector of percentage values to use for pairwise subsetting with ancestry column name vector. Only ancestries with values above the percentage are kept.}

\item{...}{Variable number of additional vectors for subsetting. Looking at the column with argument name, keep only those observations with values which are elements of the argument vector.}
}
\value{
A subset of the provided data frame.
}
\description{
Subset function optimized for admixture data.
}
\examples{
# keep only observations with anc1 > 10 and anc2 > 1
subset1 <- admix_subset(data, anc = c("anc1", "anc2"), pct = c(10, 1))

# keep only observations with values "GBR" or "FRA" in column "country"
subset2 <- admix_subset(data, country = c("GBR", "FRA"))
}
