% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{admix_barplot}
\alias{admix_barplot}
\title{Admixture Data Stacked Barplot}
\usage{
admix_barplot(
  data,
  K = 2:ncol(data),
  individuals = 1,
  sortkey = NULL,
  grouping = NULL,
  palette = "default",
  names = TRUE,
  xlab = "Individuals",
  ylab = "Ancestry",
  main = "Admixture Plot",
  noclip = FALSE
)
}
\arguments{
\item{data}{Data frame containing the admixture data.}

\item{K}{Positions of the columns containing the ancestry percentages in the provided data frame; default is second to last column.}

\item{individuals}{Position of the column with the names for the x-axis; default ist the first column.}

\item{sortkey}{Name of the column containing ancestry percentages to sort the stacked barplot with.}

\item{grouping}{Name of the column by which the stacked bars are to be grouped.}

\item{palette}{Either a color palette object, or a string to use one of the predefined color palettes ("viridis", "turbo", "alternating"); default is a modified ggplot palette.}

\item{names}{Whether to show the x-axis bar labels or not; default is "TRUE".}

\item{xlab}{A label for the x-axis.}

\item{ylab}{A label for the y-axis.}

\item{main}{A main title for the plot.}

\item{noclip}{Directly draw the plot, with clipping removed from elements. Then function does not return an object; default is set to "FALSE".}
}
\value{
A ggplot object of the stacked barplot.
}
\description{
Stacked barplot optimized for admixture data.
}
\examples{
admix_barplot(data) 

# for data frame with ancestries in third to last column, 
# with x-axis names in second column
admix_barplot(data, 
              K = 3:ncol(data), 
              individuals = 2)

# grouping data by column "country", 
# and sorting each group by ancestry column "anc1"
admix_barplot(data, 
              K = 3:ncol(data), 
              grouping = "country", 
              sortkey = "anc1")

# changing color palette to "turbo" from package 'viridis', 
# and omitting bar labels
admix_barplot(data, 
              palette = "turbo", 
              names = FALSE)

# removing title and changing axis labels text
admix_barplot(data, 
              title = "", 
              xlab = "Accessions", 
              ylab = "Ancestry [\%]")
              
# directly output grouped plot with clipping removed from elements
# (useful if there are groups with a low number of observations)
admix_barplot(data, 
              grouping = "species"
              noclip = TRUE)
}
